Address typed value codec review feedback
